+++ cartographer/mapping_3d/pose_graph/constraint_builder.cc
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "cartographer/mapping_3d/sparse_pose_graph/constraint_builder.h"
+#include "cartographer/mapping_3d/pose_graph/constraint_builder.h"
 
 #include <cmath>
 #include <functional>
@@ -36,7 +36,7 @@
 
 namespace cartographer {
 namespace mapping_3d {
-namespace sparse_pose_graph {
+namespace pose_graph {
 
 ConstraintBuilder::ConstraintBuilder(
     const mapping::pose_graph::proto::ConstraintBuilderOptions& options,
@@ -308,6 +308,6 @@ void ConstraintBuilder::DeleteScanMatcher(const mapping::SubmapId& submap_id) {
   submap_scan_matchers_.erase(submap_id);
 }
 
-}  // namespace sparse_pose_graph
+}  // namespace pose_graph
 }  // namespace mapping_3d
 }  // namespace cartographer
